Fixed calculating for nSymmetries = 0

+++ ConstantGridSolver.h
@@ -64,7 +64,7 @@ public:
     /*!
      * \brief Saves S matrix (Im and Re part separately)
      */
-    void saveS(const arma::cx_mat &S, const std::string S_type, const double E, const std::string directory);
+    void saveS(const arma::cx_mat &S, const std::string S_type, const int E, const std::string directory);
 
     void setParameters(const Parameters &parameters){
         params = parameters;
